+++ src/main/java/org/opensearch/knn/index/util/KNNEngine.java
@@ -32,6 +32,7 @@ public enum KNNEngine implements KNNLibrary {
     public static final KNNEngine DEFAULT = NMSLIB;
 
     private static final Set<KNNEngine> CUSTOM_SEGMENT_FILE_ENGINES = ImmutableSet.of(KNNEngine.NMSLIB, KNNEngine.FAISS);
+    private static final Set<KNNEngine> ENGINES_SUPPORTING_FILTERS = ImmutableSet.of(KNNEngine.LUCENE, KNNEngine.FAISS);
 
     private static Map<KNNEngine, Integer> MAX_DIMENSIONS_BY_ENGINE = Map.of(
         KNNEngine.NMSLIB,
@@ -103,6 +104,10 @@ public enum KNNEngine implements KNNLibrary {
      */
     public static Set<KNNEngine> getEnginesThatCreateCustomSegmentFiles() {
         return CUSTOM_SEGMENT_FILE_ENGINES;
+    }
+
+    public static Set<KNNEngine> getEnginesThatSupportsFilters() {
+        return ENGINES_SUPPORTING_FILTERS;
     }
 
     /**